Move fs code to system/ package.
