+++ system/fs_test.go
@@ -1,16 +1,16 @@
-package guest_test
+package system_test
 
 import (
 	"testing"
 	"testing/fstest"
 
 	"github.com/stretchr/testify/require"
-	"github.com/wetware/go/guest"
+	"github.com/wetware/go/system"
 )
 
 func TestFS(t *testing.T) {
 	t.Parallel()
 
-	err := fstest.TestFS(guest.FS{})
+	err := fstest.TestFS(system.FS{})
 	require.NoError(t, err)
 }
